<rev> [1/2] Remove `date` tag from doc template: add `date` information to GraphQL.
Now, it requires a user to input `date` in the doc before posting. To improve user experience on posting a doc, we should remove it. Since the filename of every document has date information, we can extract the date and store it to GraphQL at build time.
This part adds the `date` field to GraphQL. I will post another changes to use this information instead of that from `date` tag in the document.
<issue> Refine the post template:  https://github.com/yuxiang660/little-bee-blog/issues/1